feat(schedule-details): PR01a — address review comments
- Move bare-scheduleId redirect from page component to next.config.mjs
  (matching workflow/:runId → :runId/summary pattern); delete page.tsx
- Tabs: details + runs only (remove overview/backfills/input)
- Separate config from types/constants following workflow/domain page pattern:
  - schedule-detail-tabs.config.ts: config object only, typed with
    ScheduleDetailPageTabsConfig<K> imported from types
  - schedule-detail-page-tabs/schedule-detail-page-tabs.types.ts: defines
    ScheduleDetailTabConfig, ScheduleDetailPageTabsConfig<K>,
    ScheduleDetailTabName = keyof typeof config (mirrors WorkflowPageTabsConfig),
    and ScheduleDetailPageTabsParams
  - schedule-detail-page.types.ts: layout params only (domain/cluster/scheduleId,
    no scheduleTab) matching WorkflowPageParams
  - schedule-detail-page-tab-content.types.ts: Props uses ScheduleDetailPageTabsParams
