+++ src/views/domain-schedules/schedule-detail-page-tab-content/__tests__/schedule-detail-page-tab-content.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { render, screen } from '@/test-utils/rtl';
 
-import { type ScheduleDetailPageParams } from '../../schedule-detail-page/schedule-detail-page.types';
+import { type ScheduleDetailPageTabsParams } from '../../schedule-detail-page-tabs/schedule-detail-page-tabs.types';
 import ScheduleDetailPageTabContent from '../schedule-detail-page-tab-content';
 
 jest.mock('next/navigation', () => ({
@@ -13,19 +13,9 @@ jest.mock('next/navigation', () => ({
 }));
 
 describe(ScheduleDetailPageTabContent.name, () => {
-  it('renders placeholder for overview tab', () => {
-    setup({ scheduleTab: 'overview' });
-    expect(screen.getByText(/Overview/i)).toBeInTheDocument();
-  });
-
-  it('renders placeholder for backfills tab', () => {
-    setup({ scheduleTab: 'backfills' });
-    expect(screen.getByText(/Backfills/i)).toBeInTheDocument();
-  });
-
-  it('renders placeholder for input tab', () => {
-    setup({ scheduleTab: 'input' });
-    expect(screen.getByText(/Input/i)).toBeInTheDocument();
+  it('renders placeholder for details tab', () => {
+    setup({ scheduleTab: 'details' });
+    expect(screen.getByText(/Details/i)).toBeInTheDocument();
   });
 
   it('renders placeholder for runs tab', () => {
@@ -37,7 +27,8 @@ describe(ScheduleDetailPageTabContent.name, () => {
     const { notFound } = jest.requireMock('next/navigation');
     expect(() =>
       setup({
-        scheduleTab: 'unknown-tab' as ScheduleDetailPageParams['scheduleTab'],
+        scheduleTab:
+          'unknown-tab' as ScheduleDetailPageTabsParams['scheduleTab'],
       })
     ).toThrow('NEXT_NOT_FOUND');
     expect(notFound).toHaveBeenCalled();
@@ -45,9 +36,9 @@ describe(ScheduleDetailPageTabContent.name, () => {
 });
 
 function setup({
-  scheduleTab = 'overview',
+  scheduleTab = 'details',
 }: {
-  scheduleTab?: ScheduleDetailPageParams['scheduleTab'];
+  scheduleTab?: ScheduleDetailPageTabsParams['scheduleTab'];
 } = {}) {
   render(
     <ScheduleDetailPageTabContent
