+++ src/views/domain-schedules/schedule-detail-page-tab-content/schedule-detail-page-tab-content.tsx
@@ -7,14 +7,14 @@ import useStyletronClasses from '@/hooks/use-styletron-classes';
 import decodeUrlParams from '@/utils/decode-url-params';
 
 import scheduleDetailTabsConfig from '../schedule-detail-page/schedule-detail-tabs.config';
-import { type ScheduleDetailPageParams } from '../schedule-detail-page/schedule-detail-page.types';
+import { type ScheduleDetailPageTabsParams } from '../schedule-detail-page-tabs/schedule-detail-page-tabs.types';
 
 import { cssStyles } from './schedule-detail-page-tab-content.styles';
 import { type Props } from './schedule-detail-page-tab-content.types';
 
 export default function ScheduleDetailPageTabContent({ params }: Props) {
   const { cls } = useStyletronClasses(cssStyles);
-  const decodedParams = decodeUrlParams(params) as ScheduleDetailPageParams;
+  const decodedParams = decodeUrlParams(params) as ScheduleDetailPageTabsParams;
   const tabConfig = scheduleDetailTabsConfig[decodedParams.scheduleTab];
 
   if (!tabConfig) {
