+++ src/views/domain-schedules/schedule-detail-page/schedule-detail-page.types.ts
@@ -1,15 +1,12 @@
 import type React from 'react';
 
-import { type ScheduleDetailTabName } from './schedule-detail-tabs.config';
-
-export type ScheduleDetailPageParams = {
+export type ScheduleDetailPageLayoutParams = {
   domain: string;
   cluster: string;
   scheduleId: string;
-  scheduleTab: ScheduleDetailTabName;
 };
 
 export type Props = {
-  params: ScheduleDetailPageParams;
+  params: ScheduleDetailPageLayoutParams;
   children: React.ReactNode;
 };
